pedantic
git-svn-id: svn+ssh://oberon/svn/OMiSCID-Dev@1021 c4bd1236-c4f7-0310-a2fe-a60683d9aa4a

+++ ServiceControl/DnsSdService.cpp
@@ -61,7 +61,7 @@ bool DnsSdService::CheckProtocol( const SimpleString Protocol )
 	{
 		ProtocolLenght++;
 
-		if( ProtocolLenght >= Protocol.GetLength() )
+		if( ProtocolLenght >= (int)Protocol.GetLength() )
 			break;
 
 		if( Protocol[ProtocolLenght] == '_' )
